third_party/WebKit/LayoutTests/external/wpt/css/css-overflow/overflow-shorthand-001-expected.txt

This is a testharness.js-based test.
PASS overflow: auto auto works
FAIL overflow: hidden auto works assert_equals: expected "hidden auto" but got "auto hidden"
FAIL overflow: scroll auto works assert_equals: expected "scroll auto" but got "auto scroll"
FAIL overflow: visible auto works assert_equals: expected "visible auto" but got "auto visible"
FAIL overflow: auto hidden works assert_equals: expected "auto hidden" but got "hidden auto"
PASS overflow: hidden hidden works
FAIL overflow: scroll hidden works assert_equals: expected "scroll hidden" but got "hidden scroll"
FAIL overflow: visible hidden works assert_equals: expected "visible hidden" but got "hidden visible"
FAIL overflow: auto scroll works assert_equals: expected "auto scroll" but got "scroll auto"
FAIL overflow: hidden scroll works assert_equals: expected "hidden scroll" but got "scroll hidden"
PASS overflow: scroll scroll works
FAIL overflow: visible scroll works assert_equals: expected "visible scroll" but got "scroll visible"
FAIL overflow: auto visible works assert_equals: expected "auto visible" but got "visible auto"
FAIL overflow: hidden visible works assert_equals: expected "hidden visible" but got "visible hidden"
FAIL overflow: scroll visible works assert_equals: expected "scroll visible" but got "visible scroll"
PASS overflow: visible visible works
Harness: the test ran to completion.
